Adding in the WP Edit function
