+++ src2/save.js
@@ -20,13 +20,7 @@ import { useBlockProps } from '@wordpress/block-editor';
 export default function save( { _ref } ) {
 
 	let {
-		attributes,
-		isSelected,
-		setAttributes,
-		mergeBlocks,
-		onReplace,
-		context,
-		clientId
+		attributes
 	} = _ref;
 
 	const {
